+++ networks/bitcoin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rosen-network/bitcoin
 
+## 2.4.9
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-network/base@0.5.5
+
 ## 2.4.8
 
 ### Patch Changes