+++ networks/ethereum/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rosen-network/ethereum
 
+## 0.4.10
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-network/base@0.5.5
+
 ## 0.4.9
 
 ### Patch Changes